Update js, convert to markdown, move css to website css, add comments to message addition (js).
