+++ src/wmnds/patterns/cookies/_cookies.scss
@@ -18,7 +18,3 @@
     }
   }
 }
-
-.wmnds-website-code-example__njk .wmnds-cookies-banner {
-  display: block;
-}
