+++ src/wmnds/patterns/cookies/cookies-banner/_example.njk
@@ -1,0 +1,7 @@
+{% raw %}
+{% from "wmnds/patterns/cookies/cookies-banner/_cookies-banner.njk" import wmndsCookiesBanner %}
+
+{{
+   wmndsCookiesBanner()
+}}
+{% endraw %}
