+++ src/wmnds/patterns/cookies/cookies-manager/_example.njk
@@ -1,0 +1,10 @@
+{% raw %}
+{% from "wmnds/patterns/cookies/cookies-manager/_cookies-manager.njk" import wmndsCookiesManager %}
+
+{{
+   wmndsCookiesManager({
+      showSuccessMessage: true
+    })
+}}
+
+{% endraw %}
