+++ src/wmnds/patterns/cookies/cookies-manager/_properties.njk.json
@@ -1,0 +1,9 @@
+{
+  "wmndsCookiesManagerProps": [
+    {
+      "name": "showSuccessMessage",
+      "type": "boolean",
+      "description": "If true, it will display a success message. Defaults to <code class='wmnds-website-inline-code'>false</code>."
+    }
+  ]
+}
