+++ src/www/_partials/component-example/_component-example.scss
@@ -43,7 +43,13 @@ linear-gradient(-45deg, transparent 75%, get-color(background, 20, dark) 75%);
     max-width: 100%;
     resize: both;
     background: get-color(background);
+
+    .wmnds-cookies-banner {
+      display: block;
+    }
+
   }
+
 
   pre {
     margin: 0;
